requirements.txt
anthropic
langchain-anthropic
langchain-core
langchain-community
langgraph
streamlit
python-dotenv
chromadb
sentence-transformers